// src/frontend/src/api/generated/model/getApiV1CatalogProductsParams.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * GoldPC API
 * Объединённая OpenAPI-спецификация всех микросервисов GoldPC.
 * OpenAPI spec version: v1
 */

export type GetApiV1CatalogProductsParams = {
Category?: string;
CategoryId?: string;
ManufacturerId?: string;
ManufacturerIds?: string[];
priceMin?: number;
priceMax?: number;
InStock?: boolean;
IsFeatured?: boolean;
IsActive?: boolean;
Search?: string;
Specifications?: {[key: string]: string};
SpecificationRanges?: {[key: string]: string};
Page?: number;
PageSize?: number;
SortBy?: string;
SortOrder?: string;
};